Swap rxn rates

+++ src/scnnoise/scnnoise.cpp
@@ -929,7 +929,7 @@ namespace ScnnoiseInterface {
         this->count_save_file = new_filepath;
     }
 
-    void scNNoiSE::swap_rxn_rates (std::map<std::string, std::map<std::string, int>> rxn_rates){
+    void scNNoiSE::swap_rxn_rates (std::map<std::string, std::map<std::string, double>> rxn_rates){
       for(int gene = 0; gene<reactions.size(); ++gene){
           for(auto it = reactions[gene].rxn_rates.begin(); it != reactions[gene].rxn_rates.end(); ++it){
             it->second = rxn_rates[reactions[gene].gene_name][it->first];
